chore: WishListModel 파일 분리 (ios-h/airbnb#33)
