+++ iOS/airbnb/Present/WishList/Model/WishListItem.swift
@@ -1,0 +1,17 @@
+//
+//  WishListItem.swift
+//  airbnb
+//
+//  Created by 안상희 on 2022/06/08.
+//
+
+import Foundation
+import UIKit.UIImage
+
+struct WishListItem {
+    let image: UIImage?
+    let rating: Double?
+    let reviewCount: Int?
+    let accomodationName: String
+    let price: Int
+}
